Refactored a bit and disabled CGO.
